backend/apps/liive-auth-api/pyproject.toml
[project]
name = "liive-auth-api"
version = "0.1.0"
description = "Authentication API service for Liive"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt==4.0.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "types-python-jose>=3.3.4.20240106",
    "types-passlib>=1.7.7.20240106",
    "sqlalchemy[mypy]>=2.0.25",
    "types-sqlalchemy>=1.4.53.38",
    "liive-sql-models>=0.1.0",
    "email-validator>=2.1.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.setuptools]
py-typed = true

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
follow_imports = "normal"
warn_redundant_casts = true
warn_unused_ignores = true
disallow_any_generics = true
check_untyped_defs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.hatch.build.targets.wheel]
packages = ["src/liive_auth_api"] 